Added room types, initial stockpile and hauling jobs, dock fixture, and more
Added Room types and requirement checking to determine the type of a room e.g. based on fixtures in the room. Added zones, and first implementation of a stockpile zone, along with hauling jobs that are created for stockpiles. Added a desk fixture for dock type rooms, that needs a new type of job to function.